requirements.txt
pydub==0.23.1
numpy==1.16.2
palettable==3.1.1
scipy==1.2.1
matplotlib==3.0.3

# For pyalsaaudio you will need to install python-alsaaudio and libasound2-dev using your package manager 
pyalsaaudio==0.8.4

# For PySDL2 you'll have to install python3-sdl2 via your package manager
PySDL2==0.9.6

# You may also need tkinter (not sure about that, but if you need, here it is:
# sudo apt-get install python3-pil.imagetk
